# Manager-feeds/manager-server/swagger.openapi.yaml
openapi: 3.0.3
info:
  title: Feeds Manager Server API
  version: 1.0.0
  description: Complete OpenAPI documentation for all mounted manager-server routes.
servers:
  - url: http://localhost:3001
...
components:
  securitySchemes:
    managerCookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
    ErrorResponse:
      type: object
      required: [success, statusCode, message]
      properties:
        success:
          type: boolean
          example: false
        statusCode:
          type: integer
          example: 403
        message:
          type: string
          example: Access denied for this module
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      required: [success, msg]
      properties:
        success:
          type: boolean
          example: true
        msg:
          type: string
          example: Successfully extracted manager
    MessageResponse:
      type: object
      required: [success, msg]
      properties:
        success:
          type: boolean
          example: true
        msg:
          type: string
    AdminUser:
      type: object
      properties:
        _id:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager]
        managerType:
          type: string
          enum: [users, posts, feedback and revenue]
        status:
          type: string
          enum: [active, suspended]
        twoFactorEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: true
    AuthStatusResponse:
      type: object
      required: [isAuthenticated, user]
      properties:
        isAuthenticated:
          type: boolean
          example: true
        user:
          $ref: '#/components/schemas/AdminUser'
    FollowerEntry:
      type: object
      properties:
        username:
          type: string
      additionalProperties: true
    UserSummary:
      type: object
      properties:
        _id:
          type: string
        fullName:
          type: string
        username:
          type: string
        display_name:
          type: string
        email:
          type: string
        phone:
          type: string
        dob:
          type: string
          format: date-time
        visibility:
          type: string
          enum: [Public, Private]
        profilePicture:
          type: string
        bio:
          type: string
        type:
          type: string
          enum: [Kids, Normal, Admin]
        followers:
          type: array
          items:
            $ref: '#/components/schemas/FollowerEntry'
        followings:
          type: array
          items:
            $ref: '#/components/schemas/FollowerEntry'
        blockedUsers:
          type: array
          items:
            type: string
        isPremium:
          type: boolean
        coins:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: true
    ChannelSummary:
      type: object
      properties:
        _id:
          type: string
        channelName:
          type: string
        channelDescription:
          type: string
        channelCategory:
          type: array
          items:
            type: string
        channelLogo:
          type: string
        channelAdmin:
          type: string
        channelMembers:
          type: array
          items:
            type: string
        links:
          type: array
          items:
            type: string
        postIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: true
...
    PostPreview:
      type: object
      properties:
        id:
          type: string
        author:
          type: string
        type:
          type: string
          enum: [Reels, Img]
        url:
          type: string
        content:
          type: string
      additionalProperties: true
...
    Report:
      type: object
      properties:
        _id:
          type: string
        report_id:
          type: integer
          enum: [1, 2, 3, 4, 5, 6]
        post_id:
          type: string
        report_number:
          type: integer
        user_reported:
          type: string
        status:
          type: string
          example: Pending
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: true
    ReportWithPreview:
      allOf:
        - $ref: '#/components/schemas/Report'
        - type: object
          properties:
            scopeType:
              type: string
              enum:
                - normal_or_kids_post
                - channel_post
                - normal_chat
                - channel_chat
                - normal_or_kids_account
                - channel_account
                - unknown
            postPreview:
              oneOf:
                - $ref: '#/components/schemas/PostPreview'
                - type: 'null'
    ReachPoint:
      type: object
      properties:
        _id:
          type: object
          properties:
            year:
              type: integer
            month:
              type: integer
          additionalProperties: false
        count:
          type: integer
      additionalProperties: true
    UpdateReportStatusRequest:
      type: object
      required: [reportId, status]
      properties:
        reportId:
          type: string
        status:
          type: string
          enum: [Pending, Resolved]
...
    HomeResponse:
      type: object
      required: [success, msg]
      properties:
        success:
          type: boolean
          example: true
        msg:
          type: string
          example: Home page came
    UsersListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
...
    DashboardChannelsResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/ChannelSummary'
    RevenueResponse:
      type: object
      required: [success, rev]
      properties:
        success:
          type: boolean
          example: true
        rev:
          type: number
          example: 19450
    CountResponse:
      type: object
      required: [success, count]
      properties:
        success:
          type: boolean
          example: true
        count:
          type: integer
          example: 125
    ReachResponse:
      type: object
      required: [success, data, count]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReachPoint'
        count:
          type: integer
    ReportDataResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required: [total, pending, resolvedToday]
          properties:
            total:
              type: integer
            pending:
              type: integer
            resolvedToday:
              type: integer
    ContentActivityResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required: [postsToday, reelsToday, storiesToday]
          properties:
            postsToday:
              type: integer
            reelsToday:
              type: integer
            storiesToday:
              type: integer
    ReportsListResponse:
      type: object
      required: [success, reports]
      properties:
        success:
          type: boolean
          example: true
        reports:
          type: array
          items:
            $ref: '#/components/schemas/ReportWithPreview'
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login as manager
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Sets auth_token cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not a manager or suspended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/status:
    get:
      tags: [Auth]
      summary: Check manager authentication status
      security:
        - managerCookieAuth: []
      responses:
        '200':
          description: Session is authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStatusResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not authorized for manager portal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout current manager
      security:
        - managerCookieAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /home:
    get:
      tags: [Dashboard]
      summary: Home route
      security:
        - managerCookieAuth: []
      responses:
        '200':
          description: Home route response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /home/getUsers:
    get:
      tags: [Dashboard]
      summary: Get top users for dashboard
      description: Accessible only for manager type users.
      security:
        - managerCookieAuth: []
      responses:
        '200':
          description: Users payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /home/getChannels:
    get:
      tags: [Dashboard]
      summary: Get top channels for dashboard
      description: Accessible only for manager type users.
      security:
        - managerCookieAuth: []
      responses:
        '200':
          description: Channels payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardChannelsResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /home/getRevenue:
    get:
      tags: [Dashboard]
      summary: Get total revenue
      description: Accessible only for manager type feedback and revenue.
      security:
        - managerCookieAuth: []
      responses:
        '200':
          description: Revenue aggregate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevenueResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /home/getUserCount:
    get:
      tags: [Dashboard]
      summary: Get total users plus channels count
      description: Accessible only for manager type users.
      security:
        - managerCookieAuth: []
      responses:
        '200':
          description: Total count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /home/getReach:
    get:
      tags: [Dashboard]
      summary: Get monthly content activity reach
      description: Accessible only for manager type posts.
      security:
        - managerCookieAuth: []
      responses:
        '200':
          description: Reach metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReachResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /home/reportData:
    get:
      tags: [Dashboard]
      summary: Get report dashboard metrics
      description: Accessible only for manager type posts.
      security:
        - managerCookieAuth: []
      responses:
        '200':
          description: Reports dashboard summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDataResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /home/contentActivityToday:
    get:
      tags: [Dashboard]
      summary: Get today's posts, reels, and stories count
      description: Accessible only for manager type posts.
      security:
        - managerCookieAuth: []
      responses:
        '200':
          description: Today's activity summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentActivityResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /report/list:
    get:
      tags: [Reports]
      summary: Get reports queue for posts manager
      security:
        - managerCookieAuth: []
      responses:
        '200':
          description: Reports with post preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportsListResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /report/updateReportStatus:
    post:
      tags: [Reports]
      summary: Update report status and write manager action log
      security:
        - managerCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReportStatusRequest'
      responses:
        '200':
          description: Status updated or already same
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Scope denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...